requirements.txt
appdirs==1.4.0
Django==1.10.4
django-registration-redux==1.4
djangorestframework==3.5.3
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
